+++ CPP/type_defs.h
@@ -18,8 +18,8 @@ typedef struct run_params{
 	int         J_2N_max;
 	int         Nphi;
 	int         Nx;
-	int 		chebyshev_t;
-	int 		chebyshev_s;
+	double 		chebyshev_t;
+	double 		chebyshev_s;
 	int         Np_per_WP;
 	int         Nq_per_WP;
 	int			channel_idx;
